requirements.txt
streamlit
numpy
opencv-python-headless
scikit-image
Pillow
requests
